do logging in a separate service and make ping ingestion service to create a singleton instance
reason:
Previously I think a new logger was created
for every instance of ping controller.
And I also think ping controller gets
created for every request.
And since ping service was a class
with BindingScope.TRANSIENT scope, I think it
was also created for every controller. So,
concluding on this what I think at a high
level is - a lot of ping controllers
were created (this is okay and expected I think)
- one for every request, and for each controller
there's a new logger created and injected and for
each controller a new ping service is created
and for each ping service a logger was injected
but not sure if it was a brand new logger everytime or
not. So, to reduce the number of loggers to be created,
we have compartmentalized and put a wrapper around the
winston logger for loopback. This wrapper is implemented
as a logging service in loopback and it's only a singleton - so
only one instance of it and in it we inject one logger.
We also create only a single instance of the ping service
as we don't need new instances for every ping controller.
So, ping service is also a single.
We then use the logging service in ping
controller and ping service by injecting it
using loopback dependency injection feature

this gets rid of all the "memory leak" warnings we get when we
run tests, or when we run the service and hit it with high traffic